pyproject.toml
[project]
name = "lumiwink"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "paho-mqtt>=2.1.0",
]

[dependency-groups]
dev = [
    "mypy>=1.17.1",
    "pylance>=0.35.0",
    "pytest>=8.4.2",
]
